% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exim.r
\name{exim}
\alias{exim}
\title{Find cover of extinct or immigrant species over time}
\usage{
exim(usecover = TRUE, immigration = TRUE)
}
\arguments{
\item{usecover}{use cover of no subplots}

\item{immigration}{use immigration or extinction}
}
\description{
Find cover of extinct or immigrant species over time
}
\details{
Immigrants are species no present in 2009. Extinct are not present in 2013.
}

