% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find.origin.control.r
\name{find.origin.control}
\alias{find.origin.control}
\title{Find origin control}
\usage{
find.origin.control(tu)
}
\arguments{
\item{tu}{turfID}
}
\value{
character vector with TT1 and TTC controls
}
\description{
Find turfID of controls in origin block
}

