% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.distmoved.r
\name{plot.distmoved}
\alias{plot.distmoved}
\title{Boxplots of distance moved}
\usage{
\method{plot}{distmoved}(x, temperature = TRUE, ...)
}
\arguments{
\item{x}{output of make.dists() cbinded onto turfs.}

\item{temperature}{Split the data by temperature or precipitation}

\item{\dots}{currently unused}
}
\description{
Boxplots of distance moved
}

