% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.multitrait.r
\name{plot.multitrait}
\alias{plot.multitrait}
\title{plots multitrait result}
\usage{
\method{plot}{multitrait}(x, ...)
}
\arguments{
\item{x}{multitrait object}

\item{\dots}{extra arguments}
}
\description{
plots multitrait result
}

