% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traitselector.r
\name{traitselector}
\alias{traitselector}
\title{Find cover of species with trait}
\usage{
traitselector(usecover = TRUE, keep)
}
\arguments{
\item{usecover}{Use cover or no subplots.}

\item{keep}{logical vector of which species to include}
}
\description{
Find cover of species with trait
}

