% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/turfsppSubsets.r
\name{turfsppSubsets}
\alias{turfsppSubsets}
\title{Turf Species Subsets}
\usage{
turfsppSubsets(tu, year = 2009:2013, TT1 = TRUE, TTC = TRUE)
}
\arguments{
\item{tu}{turfID}

\item{year}{Years to use when compiling destination community}

\item{TT1}{Use TT1 when compiling target community}

\item{TTC}{Use TTC when compiling target community}
}
\value{
list of the different subsets of the community
\itemize{
 \item{origin Original species list}
 \item{final Final species list}
 \item{destination Destination block controls' species list}
 \item{potentialExtinctions Species initially present not found in destination controls}
 \item{realisedExtinctions Species that become extinct}
 \item{expectedExtinctions Species that become extinct and were potential extinctions}
 \item{unexpectedExtinctions Species that become extinct and were not potential extinctions}
 \item{persistant Species that persist}
 \item{potentialImmigration Species present in the destination controls not present in the original turf}
 \item{realisedImmigration Species that immigrate}
 \item{expectedImmigration Species that immigrate and were potentialImmigration}
 \item{unexpectedImmigration Species that immigrate and were not potentialImmigration}
}
}
\description{
Subsets of the community by their immigration and extinction status
}

